requirements.txt
asgiref==3.7.2
astroid==3.0.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.4
cssbeautifier==1.14.9
defusedxml==0.7.1
dill==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-crispy-forms==2.0
django-environ==0.11.2
django-taggit==4.0.0
django-tinymce==3.6.1
django-tree-queries==0.15.0
Django-Verify-Email==2.0.3
djlint==1.34.0
EditorConfig==0.12.3
factory-boy==3.3.0
Faker==19.6.2
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
isort==5.13.2
jsbeautifier==1.14.9
json5==0.9.14
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==4.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
python3-openid==3.2.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
